分离水平卷积时不进行饱和转换，而是在最后垂直卷积时进行，保留精度

+++ src/kernels/gaussian-blur/gaussian_blur.cu
@@ -272,8 +272,7 @@ __global__ void gaussian_blur_sep_h_kernel(const T *src, float *tmp, const WT *w
             const WT  weight = weights_x[w_x];
             sum += weight * src[(y * img_w + xx) * CH + c];
         }
-        // 写入中间结果；对整数类型会产生一次中间舍入/饱和
-        tmp[base + c] = saturate_cast<float>(sum);
+        tmp[base + c] = sum;
     }
 }
